rag/requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-vector-stores-qdrant
qdrant-client[local]
fastapi
uvicorn
python-dotenv
httpx
